Rendering Core, Direct X added
This took a while, mostly because Direct X and setting it up with a
project is complicated. But the rendering core folder is added, and
Direct X is set up. Remember standard procedure is to delete the "ipch"
folders, "Debug" folders, and .sdb files. This shrinks total file size
tremendously without breaking anything.